+++ doc/models.rst
@@ -52,7 +52,7 @@ For fixed parameters the ``"init"`` value gives the value of that parameter to u
 Nearly all parameters used by FSPS can be set (or varied) here.
 When fitting galaxies the default FSPS parameter values will be used unless specified in a fixed parameter,
 e.g. ``imf_type`` can be changed by including it as a fixed parameter with value given by ``"init"``.
-More generally any parameter used by the source basis object to build an SED can be in the ``model_params`` list.
+More generally any parameter used by the ``sources`` object to build an SED can be in the ``model_params`` list.
 
 
 The ``load_model()`` method
@@ -66,9 +66,9 @@ so the model can be modified based on keywords given there (or at the command li
 The ``load_sps()`` function
 -------------------------------------
 
-The likelihood function and SED models take an object (``sps``) from  ``prospect.source_basis`` as an argument.
+The likelihood function and SED models take an object (``sps``) from  ``prospect.sources`` as an argument.
 This object should be returned by the ``load_sps()`` function in the **parameter file**.
-The source basis object generally includes all the spectral libraries necessary to build a model,
+The ``sps`` object generally includes all the spectral libraries necessary to build a model,
 as well as some model building code.
 This object is defined globally to enable multiprocessing, since generally it can't (or shouldn't) be serialized
 and sent to other processors.
