+++ misc/timings_pyfsps.py
@@ -4,7 +4,7 @@ import time, os, subprocess, re, sys
 import numpy as np
 #import matplotlib.pyplot as pl
 import fsps
-from prospect import source_basis as sps_basis
+from prospect import sources as sps_basis
 from prospect.models import sedmodel
