+++ prospect/__init__.py
@@ -1,6 +1,6 @@
 from . import models
 from . import fitting
 from . import io
-from . import source_basis
+from . import sources
 from . import utils
 from . import likelihood
