--- src/extra_parameters.tex
@@ -1,4 +1,0 @@
-REAL :: logzsol
-REAL :: ncomp
-DIMENSION (ncomp) :: ssp_amplitudes, age_index, zindex
-
