--- src/template.f90
@@ -1,19 +1,0 @@
-SUBROUTINE SOMETHING(in,out)
-  !This does something
-  !
-  ! Inputs
-  ! ------
-  !
-  ! in [integer]:
-  !   The input
-  !
-  ! Outputs
-  ! -------
-  !
-  ! out [double precision (ndim)]:
-  !   The output.
-
-  USE blah
-  IMPLICIT NONE
-
-END SUBROUTINE
